requirements.txt
halo
python-nmap
scapy
colorama
argparse
netifaces
random2
subprocess.run
jsons
tabulate
TIME-python